[bees] Extract Python app parts to `packages/bees/app`
Extracted the Python files belonging to the reference application and CLI tools from `packages/bees/bees` into a new sibling package `packages/bees/app`, and updated associated scripts and tests.

To improve the layering of the `packages/bees` package, separating the core library logic from the application entry points and CLI tools.

- Created `app/` directory and initialized it as a Python package.
- Moved `server.py`, `cli.py`, `add_ticket.py`, `drain.py`, `edit_tags.py`, `respond.py`, and `run_playbook.py` from `bees/` to `app/`.
- Moved and updated `bees/__main__.py` to `app/__main__.py` to point to `app.cli`.
- Updated `package.json` scripts to use `python -m app.xxx` instead of `python -m bees.xxx`.
- Updated `app/server.py` to fix hardcoded `bees.server` in `uvicorn.run`.
- Updated `tests/test_query.py` to import from `app.server`.

- Ran `npm run test:python` and all 161 tests passed.
- Verified that `npm run ticket:add` works with the new layout.
