+++ packages/bees/app/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2026 Google LLC
+# SPDX-License-Identifier: Apache-2.0
