+++ packages/bees/app/__main__.py
@@ -1,8 +1,8 @@
 # Copyright 2026 Google LLC
 # SPDX-License-Identifier: Apache-2.0
 
-"""Allow ``python -m bees.cli``."""
+"""Allow ``python -m app``."""
 
-from bees.cli import main
+from app.cli import main
 
 main()
