+++ packages/bees/app/server.py
@@ -585,7 +585,7 @@ async def events() -> EventSourceResponse:
 def main() -> None:
     import uvicorn
     uvicorn.run(
-        "bees.server:app",
+        "app.server:app",
         host="0.0.0.0",
         port=3200,
         reload=True,
