+++ packages/bees/package.json
@@ -5,14 +5,14 @@
   "description": "Bees — agent swarm orchestration framework",
   "scripts": {
     "setup": "wireit",
-    "session:start": ".venv/bin/python -m bees.cli",
-    "ticket:add": ".venv/bin/python -m bees.add_ticket",
-    "ticket:drain": ".venv/bin/python -m bees.drain",
-    "ticket:respond": ".venv/bin/python -m bees.respond",
-    "playbook:run": ".venv/bin/python -m bees.run_playbook",
+    "session:start": ".venv/bin/python -m app.cli",
+    "ticket:add": ".venv/bin/python -m app.add_ticket",
+    "ticket:drain": ".venv/bin/python -m app.drain",
+    "ticket:respond": ".venv/bin/python -m app.respond",
+    "playbook:run": ".venv/bin/python -m app.run_playbook",
     "dev": "npm run dev:server & npm run dev:web",
     "dev:clean": "rm -rf hive/tickets hive/logs && npm run dev",
-    "dev:server": ".venv/bin/python -m bees.server",
+    "dev:server": ".venv/bin/python -m app.server",
     "dev:web": "cd web && npm run dev",
     "dev:hivetool": "cd hivetool && npm run dev",
     "test:python": ".venv/bin/python -m pytest tests/ -v"
